fix #4612
